Added a helper class, and implemented exploratory analysis of the data.
